sharing template in _app got button to be purple themed renamed home-component to form-component
