+++ components/Message.js
@@ -1,4 +1,3 @@
-import React from "react";
 import Container from "@material-ui/core/Container";
 import Typography from "@material-ui/core/Typography";
