+++ pages/index.js
@@ -1,31 +1,13 @@
-import { useEffect } from "react";
-import Head from "next/head";
+import _ from "lodash";
 import appPageHandler from "middleware/app-page-handler";
-import FooterComponent from "components/footer-component";
-import HeaderComponent from "components/header-component";
 import StepsComponent from "components/steps-component";
-import HomeComponent from "components/home-component";
+import FormComponent from "components/form-component";
 
 export default function Index(props) {
-  useEffect(() => {
-    // this is required for MUI
-    // Remove the server-side injected CSS.
-    const jssStyles = document.querySelector("#jss-server-side");
-    if (jssStyles) {
-      jssStyles.parentElement.removeChild(jssStyles);
-    }
-  }, []);
-
   return (
     <>
-      <Head>
-        <title>ShareSecure</title>
-      </Head>
-
-      <HeaderComponent {...props} />
-      <HomeComponent {...props} />
+      <FormComponent {...props} />
       <StepsComponent {...props} />
-      <FooterComponent {...props} />
     </>
   );
 }
@@ -33,7 +15,7 @@ export default function Index(props) {
 export function getServerSideProps(ctx) {
   // middleware
   appPageHandler(ctx.req, ctx.res);
-  const appConfig = ctx.req.appConfig;
+  const appConfig = _.get(ctx, "req.appConfig", {});
 
   // pass config data to page props
   return { props: { ...appConfig } };
