+++ svg/logo.js
@@ -1,4 +1,3 @@
-import { React } from "react";
 
 const Logo = ({ width = "52", height = "52" }) => {
   return (
